processor_service/requirements.txt
python-dotenv
paramiko
